Improved: The updater source code, Fixed: An issue with interfaces detection (see more in commit description)
Refined: Code.
Improved: The updater's source code now incorporates the capability to detect the availability of a new update not only through versioning but also via date and time.
Fixed: An issue where an interface could lack an IP address, and an error ("Developer didn't expect this scenario to be possible") would be raised . Now, such instances will be skipped.